+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Close button is aligned to top of modal header
 
+### ValueTable
+
+- Adds `variant` prop with values "relaxed", "standard", "condensed", "compact".
+
 ## 9.1.0
 
 ### Select